fix: navigation elements change
