+++ apps/web/components/Skeletons/PaypalButton.vue
@@ -1,0 +1,3 @@
+<template>
+  <div class="h-[55px] bg-gray-300 rounded w-full animate-pulse" />
+</template>
