+++ docs/changelog/changelog_en.md
@@ -13,6 +13,7 @@
 
 ### 🩹 Fixed
 
+- Fixed an issue where page elements changed during navigation.
 - Fixed a hydration error when fetching recommended products on the homepage.
 
 ## v1.9.0 (2025-01-23) <a href="https://github.com/plentymarkets/plentyshop-pwa/compare/v1.8.0...v1.9.0" target="_blank" rel="noopener"><b>Overview of all changes</b></a>
